100% Functional sem login também player_details.html

+++ app/views.py
@@ -209,19 +209,19 @@ def players(request):
 
 
 def player_details(request, id):
-    # if request.method == 'POST':
-    # FavouritePlayer.objects.filter(player_id=id, user_id=request.user.id).delete()
-    # print("yupi")  # Not Working
-    normal=NormalUser.objects.get(user__username=request.user.username)
-    print(normal.id)
+
     if request.method == 'POST':
+        if not request.user.is_authenticated or request.user.username == 'admin':
+            return redirect('/login')
+        normal = NormalUser.objects.get(user__username=request.user.username)
+        print(normal.id)
         if 'remove' in request.POST:
             s = FavouritePlayer.objects.get(player_id=id, user_id=normal.id)
             s.delete()
             print("Removed from Favourites")
             return HttpResponseRedirect(str(id))
         elif 'add' in request.POST:
-            s=FavouritePlayer(player_id=id, user_id=normal.id)
+            s = FavouritePlayer(player_id=id, user_id=normal.id)
             s.save()
             print("Added to Favourites")
             return HttpResponseRedirect(str(id))
@@ -246,12 +246,17 @@ def player_details(request, id):
                 'formComment': MakeCommentForm()
             }
             return HttpResponseRedirect(str(id))  # render(request, 'player_details.html', t_parms)
-    if FavouritePlayer.objects.filter(player_id=id, user_id=normal.id):
-        favouriteplayer = True
-        print("Está nos Favoritos")
-    else:
+
+    if not request.user.is_authenticated or request.user.username == 'admin':
         favouriteplayer = False
-        print("Não está nos Favoritos")
+    else:
+        normal = NormalUser.objects.get(user__username=request.user.username)
+        if FavouritePlayer.objects.filter(player_id=id, user_id=normal.id):
+            favouriteplayer = True
+            print("Está nos Favoritos")
+        else:
+            favouriteplayer = False
+            print("Não está nos Favoritos")
     t_parms = {
         'player': Player.objects.get(id=id),
         'teams': set(Team.objects.filter(playerplaysfor__player_id=id).distinct()),  # TODO: Encontrar melhor solução
@@ -260,7 +265,6 @@ def player_details(request, id):
         'age': int((datetime.date.today() - Player.objects.get(id=id).birthday).days / 365),
         'favouriteplayer': favouriteplayer,
         'formComment': MakeCommentForm(),
-        'formFavourite': FavouriteForm(),
     }
     return render(request, 'player_details.html', t_parms)
